refactor: overview page upd
