--- content/docs/index.mdx
@@ -1,8 +1,0 @@
----
-title: Overview
-pageTitle: ''
----
-
-import { IndexPage } from '@/components/pages/index-page';
-
-<IndexPage />
